+++ gnani/stt/__init__.py
@@ -25,7 +25,7 @@ from gnani.stt.exceptions import (
     StreamError,
 )
 
-__version__ = "0.2.1"
+__version__ = "0.3.0"
 __all__ = [
     "SAMPLE_RATE_8K",
     "SAMPLE_RATE_16K",